Make pkg dir and move some packages
