+++ cmd/dev/cmd/delete.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/cybozu-go/log"
-	"github.com/cybozu-go/neco-gcp/functions"
-	"github.com/cybozu-go/neco-gcp/gcp"
+	"github.com/cybozu-go/neco-gcp/pkg/functions"
+	"github.com/cybozu-go/neco-gcp/pkg/gcp"
 	"github.com/cybozu-go/well"
 	"github.com/spf13/cobra"
 )
