+++ cmd/necogcp/cmd/necotest_createinstance.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/cybozu-go/log"
 	necogcp "github.com/cybozu-go/neco-gcp"
-	"github.com/cybozu-go/neco-gcp/gcp"
+	"github.com/cybozu-go/neco-gcp/pkg/gcp"
 	"github.com/cybozu-go/well"
 	"github.com/spf13/cobra"
 )
