+++ necogcp-app/main.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/cybozu-go/log"
-	"github.com/cybozu-go/neco-gcp/gcp"
-	"github.com/cybozu-go/neco-gcp/gcp/app"
+	"github.com/cybozu-go/neco-gcp/pkg/app"
+	"github.com/cybozu-go/neco-gcp/pkg/gcp"
 	"google.golang.org/appengine"
 	"sigs.k8s.io/yaml"
 )
