+++ pkg/gcp/artifacts.go
