+++ pkg/gcp/assets.go
