+++ pkg/gcp/assets/etc/containers/registries.conf
