+++ pkg/gcp/assets/etc/docker/daemon.json
