+++ pkg/gcp/assets/etc/profile.d/go.sh
