+++ pkg/gcp/compute.go
