+++ pkg/gcp/hostvm.go
