+++ pkg/gcp/image.go
