+++ pkg/gcp/types.go
