+++ pkg/slack/log/invalid/from_cloud_function.json
