+++ pkg/slack/log/invalid/operation_first.json
