+++ pkg/slack/log/valid/delete.json
