+++ pkg/slack/log/valid/startup_script.json
